Work in progress to better handle model import
