+++ ample/util/pdb_edit.py
@@ -111,135 +111,6 @@ def calpha_only(inpdb, outpdb):
         raise RuntimeError("Error stripping PDB to c-alpha atoms")
 
 
-def check_pdb_directory(directory, single=True, allsame=True, sequence=None):
-    """Check a directory of pdb files to ensure they are valid
-
-    Parameters
-    ----------
-    direcotry : str
-       Path to a directory containing PDB files
-    sequence : str
-       single-letter protein sequence - if given a check will be made that all
-       models are of this sequence
-    single : bool
-       if True check each pdb only contains a single model
-    allsame : bool
-       only extract a file if the suffix is in the list
-
-    Returns
-    -------
-    bool
-        True if models all validated
-    """
-    logger.info("Checking pdbs in directory: {0}".format(directory))
-    if not os.path.isdir(directory):
-        logger.critical("Cannot find directory: {0}".format(directory))
-        return False
-    models = glob.glob(os.path.join(directory, "*.pdb"))
-    if not len(models):
-        logger.critical("Cannot find any pdb files in directory: {0}".format(directory))
-        return False
-    if not (single or sequence or allsame):
-        return True
-    return check_pdbs(models, sequence=sequence, single=single, allsame=allsame)
-
-
-def check_pdbs(models, single=True, allsame=True, sequence=None):
-    """Check a list of PDB files to ensure they are valid
-
-    Parameters
-    ----------
-    models : list
-       A list of PDB files
-    sequence : str
-       single-letter protein sequence - if given a check will be made that all
-       models are of this sequence
-    single : bool
-       if True check each pdb only contains a single model
-    allsame : bool
-       only extract a file if the suffix is in the list
-
-    Returns
-    -------
-    bool
-        True if models all validated
-    """
-    if allsame and not sequence:
-        # Get sequence from first model
-        try:
-            h = iotbx.pdb.pdb_input(models[0]).construct_hierarchy()
-        except Exception as e:
-            s = "*** ERROR reading sequence from first pdb: {0}\n{1}".format(models[0], e)
-            logger.critical(s)
-            return False
-        sequence = _sequence1(h)  # only one model/chain
-    errors = []
-    multi = []
-    no_protein = []
-    sequence_err = []
-    unnamed_chain = []
-    for pdb in models:
-        try:
-            h = iotbx.pdb.pdb_input(pdb).construct_hierarchy()
-        except Exception as e:
-            errors.append((pdb, e))
-            continue
-        if not single:
-            continue
-        if not (h.models_size() == 1 and h.models()[0].chains_size() == 1):
-            multi.append(pdb)
-            continue
-        # single chain from one model so check is protein
-        if not h.models()[0].chains()[0].is_protein():
-            no_protein.append(pdb)
-            continue
-        if not h.models()[0].chains()[0].id.strip():  # Check we have a named chain
-            unnamed_chain.append(pdb)
-            continue
-        if sequence:
-            s = _sequence1(h)  # only one chain/model
-            if not s == sequence:
-                sequence_err.append((pdb, s))
-
-    if not (len(errors) or len(multi) or len(sequence_err) or len(no_protein) or len(unnamed_chain)):
-        logger.info("check_pdb_directory - pdb files all seem valid")
-        return True
-
-    s = "\n"
-    if len(errors):
-        s = "*** ERROR ***\n"
-        s += "The following pdb files have errors:\n\n"
-        for pdb, e in errors:
-            s += "{0}: {1}\n".format(pdb, e)
-
-    if len(multi):
-        s += "\n"
-        s += "The following pdb files have more than one chain:\n\n"
-        for pdb in multi:
-            s += "{0}\n".format(pdb)
-
-    if len(no_protein):
-        s += "\n"
-        s += "The following pdb files do not appear to contain any protein:\n\n"
-        for pdb in no_protein:
-            s += "{0}\n".format(pdb)
-
-    if len(unnamed_chain):
-        s += "\n"
-        s += "The following pdb files do not have named chains:\n\n"
-        for pdb in unnamed_chain:
-            s += "{0}\n".format(pdb)
-
-    if len(sequence_err):
-        s += "\n"
-        s += "The following pdb files have differing sequences from the reference sequence:\n\n{0}\n\n".format(sequence)
-        for pdb, seq in sequence_err:
-            s += "PDB: {0}\n{1}\n".format(pdb, seq)
-
-    logger.critical(s)
-    return False
-
-
 def chain_id_is_blank(chain):
     return (isinstance(chain.id, str) and len(chain.id.strip()) == 0)
